+++ test/react-native/features/fixtures/app/react_native_navigation_js/Scenarios.js
@@ -1,0 +1,3 @@
+// navigation.feature
+export { ReactNavigationBreadcrumbsEnabledScenario } from './scenarios/ReactNavigationBreadcrumbsEnabledScenario'
+export { ReactNavigationBreadcrumbsDisabledScenario } from './scenarios/ReactNavigationBreadcrumbsDisabledScenario'