+++ test/react-native/features/fixtures/app/react_native_navigation_js/scenarios/ReactNavigationBreadcrumbsEnabledScenario.js
@@ -1,0 +1,3 @@
+import Scenario from './Scenario'
+
+export class ReactNavigationBreadcrumbsEnabledScenario extends Scenario {}