+++ test/react-native/features/fixtures/app/react_navigation_js/install.sh
@@ -1,0 +1,9 @@
+npm i @bugsnag/plugin-react-navigation@$BUGSNAG_VERSION --registry=$REGISTRY_URL
+
+npm i @react-native-community/masked-view --registry=$REGISTRY_URL
+npm i @react-navigation/native --registry=$REGISTRY_URL
+npm i @react-navigation/stack --registry=$REGISTRY_URL
+npm i react-native-gesture-handler --registry=$REGISTRY_URL
+npm i react-native-reanimated --registry=$REGISTRY_URL
+npm i react-native-safe-area-context --registry=$REGISTRY_URL
+npm i react-native-screens --registry=$REGISTRY_URL